Initial commit. Most basic things done already
Opening, closing, reopening, modifying, name, severity level, and
description of tickets are already setup.

The ticket format is as follows, with line numbers:

1  [TICKET_NAME]
2  [SEVERITY (optional)]
3  [DESCRIPTION (optional)]
.  [DESCRIPTION (continued)]
n  [DESCRIPTION (continued)]

As suggested by the line numbers, the description can be arbitrarily
long, whereas the name and severity only take up one line. TICKET_NAME
is mandatory to create the ticket.

Right now there are only a few functions. If any more functionality is
to be added, create- and update-tickets may be moved to their own
functions to allow better code reuse.

Right now the filename is `tkt.sh', but it'd be nicer for actual use
to just be called `tkt'. The .sh extension is just to make
syntax-highlighting easier when editing.